+++ app/components/Radio/messages.js
@@ -1,8 +1,0 @@
-/*
- * Radio Messages
- *
- * This contains all the text for the Radio component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({});