+++ app/components/Select/messages.js
@@ -1,8 +1,0 @@
-/*
- * Select Messages
- *
- * This contains all the text for the Select component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({});