+++ app/containers/SidePanel/selectors.js
@@ -7,6 +7,17 @@ const getSelectedInteractionId = createSelector(
   (agentDesktop) => agentDesktop.get('selectedInteractionId')
 );
 
+const getSelectedInteractionIsVoice = createSelector(
+  selectAgentDesktopDomain,
+  (agentDesktop) => {
+    const selectedInteraction = agentDesktop.get('interactions').toJS().find(
+      (interaction) => interaction.interactionId === agentDesktop.get('selectedInteractionId')
+    );
+    return selectedInteraction !== undefined && selectedInteraction.channelType === 'voice';
+  }
+);
+
 export {
   getSelectedInteractionId,
+  getSelectedInteractionIsVoice,
 };